(PC-20410) refacto(parseSearchParameters): rename parseSearchPArameters and useParseSearchPArameters into useAdaptOffersPlaylistParameters and adaptOffersPlaylistParameters
